--- WeatherStore/WebContent/WEB-INF/web.xml
@@ -1,43 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<web-app xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns="http://java.sun.com/xml/ns/javaee" xsi:schemaLocation="http://java.sun.com/xml/ns/javaee http://java.sun.com/xml/ns/javaee/web-app_3_0.xsd" id="WebApp_ID" version="3.0">
-  <display-name>WeatherStore</display-name>
-  <welcome-file-list>
-    <welcome-file>index.html</welcome-file>
-  </welcome-file-list>
-    <!--<servlet>
-        <servlet-name>com.tia.weatherStorePack</servlet-name>
-        <servlet-class>org.glassfish.jersey.servlet.ServletContainer</servlet-class>
-        <init-param>
-            <param-name>javax.ws.rs.Application</param-name>
-            <param-value>com.tia.weatherStorePack.webapp.MyApplication</param-value>
-        </init-param>
-        <load-on-startup>1</load-on-startup>
-    </servlet>
-    <servlet-mapping>
-        <servlet-name>com.tia.weatherStorePack</servlet-name>
-        <url-pattern>/*</url-pattern>
-    </servlet-mapping>-->
-
-  <!--<servlet>
-    <servlet-name>Jersey REST Service</servlet-name>
-    <servlet-class>com.sun.jersey.spi.container.servlet.ServletContainer</servlet-class>
-    <init-param>
-      <param-name>com.sun.jersey.config.property.packages</param-name>
-      <param-value>com.tia.weatherStorePack</param-value>
-    </init-param>
-    <load-on-startup>1</load-on-startup>
-  </servlet>
-  <servlet-mapping>
-    <servlet-name>Jersey REST Service</servlet-name>
-    <url-pattern>/rest/*</url-pattern>
-  </servlet-mapping>-->
-  
-    <servlet>
-    <servlet-name>WeatherStore</servlet-name>
-    <servlet-class>com.tia.weatherStorePack.GetWeatherService</servlet-class>
-  </servlet>
-  <servlet-mapping>
-    <servlet-name>WeatherStore</servlet-name>
-    <url-pattern>/getweather</url-pattern>
-  </servlet-mapping>
-</web-app>
