--- WeatherStore/WebContent/css/styles.css
@@ -1,41 +1,0 @@
-body { background-color: #00FFCC }
-a:hover { color: red }
-.title { 
-  margin-left: auto; 
-  margin-right: auto; 
-}
-.title th { 
-  font-size: 28px;
-  font-family: Arial, Helvetica, sans-serif;
-}
-th { 
-     font-size: 18px;
-     font-family: Arial, Helvetica, sans-serif;
-}
-td { font-size: 18px;
-     font-family: Arial, Helvetica, sans-serif;
-}
-legend {
-  font-weight: bold;
-  color: black;
-  background-color: #eeeeee;
-  border: 1px solid #999999;
-  padding: 3px 2px;
-}
-.ajaxResult { color: #440000;
-              font-weight: bold;
-              font-size: 18px;
-              font-family: Arial, Helvetica, sans-serif;
-}
-h2 { color: #440000;
-	  font-weight: bold;
-	  font-size: 18px;
-	  font-family: Arial, Helvetica, sans-serif;
-}
-.error { background-color: yellow;
-	      color: red;
-	      font-weight: bold;
-	      font-size: 20px;
-	      font-family: Arial, Helvetica, sans-serif;
-	      border-style: inset;
-}
