+++ WeatherStore/build.gradle
@@ -1,24 +1,35 @@
-apply plugin: 'java'
-apply plugin: 'war'
-apply plugin: 'application'
+buildscript {
+    repositories {
+        mavenCentral()
+    }
+    dependencies {
+        classpath("org.springframework.boot:spring-boot-gradle-plugin:1.2.2.RELEASE")
+    }
+}
 
-mainClassName = 'com.tia.weatherStorePack.testMain'
+apply plugin: 'java'
+apply plugin: 'application'
+apply plugin: 'spring-boot'
+
+mainClassName = 'com.tia.weatherStorePack.TestApp'
+
 
 repositories {
     mavenCentral()
 }
 
 dependencies {
-    compile 'org.slf4j:slf4j-log4j12:1.5.6'
-    compile 'org.slf4j:slf4j-simple:1.7.10'
-    compile 'org.quartz-scheduler:quartz:2.1.5'
+    compile 'log4j:log4j:1.2.17'
+    compile 'org.apache.logging.log4j:log4j-api:2.2'
+    compile 'org.quartz-scheduler:quartz:2.2.1'
     compile 'javax:javaee-api:7.0'
     compile 'javax.servlet:javax.servlet-api:3.0.1'
     compile 'org.glassfish.jersey.containers:jersey-container-grizzly2-servlet:2.0-m07-1'
     compile 'org.glassfish.jersey.containers:jersey-container-servlet-core:2.16'
-    compile 'org.json:json:20141113'
+    compile 'org.json:json:20140107'
     compile 'org.postgresql:postgresql:9.3-1100-jdbc41'
     compile 'org.apache.tomcat:tomcat-api:7.0.59'
     compile 'net.aksingh:owm-japis:2.5.0.3'
+    compile 'org.springframework.boot:spring-boot-starter-web'
 
 }
