--- WeatherStore/deps.txt
@@ -1,18 +1,0 @@
-<dependency>
-    <groupId>org.slf4j</groupId>
-    <artifactId>slf4j-log4j12</artifactId>
-    <version>1.5.6</version>
-</dependency>
-
-<dependencies>
-	<dependency>
-		<groupId>org.quartz-scheduler</groupId>
-		<artifactId>quartz</artifactId>
-		<version>2.1.5</version>
-	</dependency>
-</dependencies>
-<dependency>
-	<groupId>javax</groupId>
-	<artifactId>javaee-api</artifactId>
-	<version>7.0</version>
-</dependency>
