--- WeatherStore/pom.xml
@@ -1,154 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!-- DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS HEADER. Copyright (c) 
-	2011-2015 Oracle and/or its affiliates. All rights reserved. The contents 
-	of this file are subject to the terms of either the GNU General Public License 
-	Version 2 only ("GPL") or the Common Development and Distribution License("CDDL") 
-	(collectively, the "License"). You may not use this file except in compliance 
-	with the License. You can obtain a copy of the License at http://glassfish.java.net/public/CDDL+GPL_1_1.html 
-	or packager/legal/LICENSE.txt. See the License for the specific language 
-	governing permissions and limitations under the License. When distributing 
-	the software, include this License Header Notice in each file and include 
-	the License file at packager/legal/LICENSE.txt. GPL Classpath Exception: 
-	Oracle designates this particular file as subject to the "Classpath" exception 
-	as provided by Oracle in the GPL Version 2 section of the License file that 
-	accompanied this code. Modifications: If applicable, add the following below 
-	the License Header, with the fields enclosed by brackets [] replaced by your 
-	own identifying information: "Portions Copyright [year] [name of copyright 
-	owner]" Contributor(s): If you wish your version of this file to be governed 
-	by only the CDDL or only the GPL Version 2, indicate your decision by adding 
-	"[Contributor] elects to include this software in this distribution under 
-	the [CDDL or GPL Version 2] license." If you don't indicate a single choice 
-	of license, a recipient has the option to distribute your version of this 
-	file under either the CDDL, the GPL Version 2 or to extend the choice of 
-	license to its licensees as provided above. However, if you add GPL Version 
-	2 code and therefore, elected the GPL Version 2 license, then the option 
-	applies only if the new code is made subject to such option by the copyright 
-	holder. -->
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-
-	<modelVersion>4.0.0</modelVersion>
-
-	<groupId>tia.weatherStore</groupId>
-	<artifactId>weather-store</artifactId>
-	<packaging>war</packaging>
-	<name>weather-store-webapp</name>
-	<version>1.0s</version>
-
-	<description>Test weather receiving app</description>
-
-	<dependencies>
-		<dependency>
-			<groupId>org.slf4j</groupId>
-			<artifactId>slf4j-log4j12</artifactId>
-			<version>1.5.6</version>
-		</dependency>
-
-		<dependency>
-			<groupId>org.slf4j</groupId>
-			<artifactId>slf4j-simple</artifactId>
-			<version>1.7.10</version>
-		</dependency>
-
-		<dependency>
-			<groupId>org.quartz-scheduler</groupId>
-			<artifactId>quartz</artifactId>
-			<version>2.1.5</version>
-		</dependency>
-
-		<dependency>
-			<groupId>javax</groupId>
-			<artifactId>javaee-api</artifactId>
-			<version>7.0</version>
-		</dependency>
-
-		<dependency>
-			<groupId>javax.servlet</groupId>
-			<artifactId>javax.servlet-api</artifactId>
-			<version>3.0.1</version>
-		</dependency>
-
-		<dependency>
-			<groupId>org.glassfish.jersey.containers</groupId>
-			<artifactId>jersey-container-grizzly2-servlet</artifactId>
-			<version>2.0-m07-1</version>
-		</dependency>
-
-		<dependency>
-			<groupId>org.glassfish.jersey.containers</groupId>
-			<artifactId>jersey-container-servlet-core</artifactId>
-			<version>2.16</version>
-		</dependency>
-
-
-		<dependency>
-			<groupId>org.json</groupId>
-			<artifactId>json</artifactId>
-			<version>20141113</version>
-		</dependency>
-
-		<dependency>
-			<groupId>org.postgresql</groupId>
-			<artifactId>postgresql</artifactId>
-			<version>9.3-1100-jdbc41</version>
-		</dependency>
-
-		<dependency>
-			<groupId>org.apache.tomcat</groupId>
-			<artifactId>tomcat-api</artifactId>
-			<version>7.0.59</version>
-		</dependency>
-
-		<dependency>
-			<groupId>net.aksingh</groupId>
-			<artifactId>owm-japis</artifactId>
-			<version>2.5.0.3</version>
-		</dependency>
-
-	</dependencies>
-
-	<build>
-		<plugins>
-			<!-- Run the application using "mvn jetty:run" to deploy to Jetty -->
-			<plugin>
-				<groupId>org.mortbay.jetty</groupId>
-				<artifactId>maven-jetty-plugin</artifactId>
-			</plugin>
-			<!-- Run the application using "mvn exec:java" to deploy to Grizzly Servlet -->
-			<plugin>
-				<groupId>org.codehaus.mojo</groupId>
-				<artifactId>exec-maven-plugin</artifactId>
-				<configuration>
-					<mainClass>com.tia.weatherStorePack.testMain</mainClass>
-				</configuration>
-			</plugin>
-			<plugin>
-				<groupId>org.apache.maven.plugins</groupId>
-				<artifactId>maven-war-plugin</artifactId>
-				<configuration>
-					<webXml>WebContent\WEB-INF\web.xml</webXml>
-					<archiveClasses>true</archiveClasses>
-				</configuration>
-			</plugin>
-		</plugins>
-	</build>
-
-	<profiles>
-		<profile>
-			<id>release</id>
-			<build>
-				<plugins>
-					<plugin>
-						<groupId>org.codehaus.mojo</groupId>
-						<artifactId>xml-maven-plugin</artifactId>
-					</plugin>
-					<plugin>
-						<groupId>org.apache.maven.plugins</groupId>
-						<artifactId>maven-assembly-plugin</artifactId>
-					</plugin>
-				</plugins>
-			</build>
-		</profile>
-	</profiles>
-
-</project>
