--- WeatherStore/src/com/tia/weatherStorePack/CityRecord.java
@@ -1,6 +1,0 @@
-package com.tia.weatherStorePack;
-
-public class CityRecord {
-	public String name;
-	public String country;
-}
