+++ WeatherStore/src/main/java/com/tia/weatherStorePack/CityRecord.java
@@ -1,0 +1,7 @@
+package com.tia.weatherStorePack;
+
+// stores description of the city for searching weather data on openweathermap.com
+public class CityRecord {
+	public String name;
+	public String country;
+}
