+++ WeatherStore/src/main/java/com/tia/weatherStorePack/JsonStrings.java
@@ -1,0 +1,7 @@
+package com.tia.weatherStorePack;
+
+public class JsonStrings {
+	public final static String JSON_TAG_CITIES = "cities";
+	public final static String JSON_TAG_CITY_NAME = "name";
+	public final static String JSON_TAG_COUNTRY = "country";
+}
