+++ WeatherStore/src/main/java/com/tia/weatherStorePack/MvcConfig.java
@@ -1,0 +1,17 @@
+package com.tia.weatherStorePack;
+
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.servlet.config.annotation.ViewControllerRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurerAdapter;
+
+@Configuration
+public class MvcConfig extends WebMvcConfigurerAdapter {
+
+    @Override
+    public void addViewControllers(ViewControllerRegistry registry) {
+        registry.addViewController("/index").setViewName("home");
+        registry.addViewController("/").setViewName("home");
+    }
+
+}
