--- WeatherStore/src/weatherStorePack/CityRecord.java
@@ -1,6 +1,0 @@
-package weatherStorePack;
-
-public class CityRecord {
-	public String name;
-	public String country;
-}
